move tests to match primary dir
