+++ tests/karma/BookNavigator/bookmarks/bookmarks-list.test.js
@@ -4,7 +4,7 @@ import {
   expect,
   oneEvent,
 } from '@open-wc/testing';
-import { IABookmarksList } from '../../../src/BookNavigator/menu-panels/bookmarks/bookmarks-list.js';
+import { IABookmarksList } from '../../../src/BookNavigator/bookmarks/bookmarks-list.js';
 
 customElements.define('ia-bookmarks-list', IABookmarksList);
